Add better handling when semantic scholar is down
This change updates how citation counts are handled when we are unable
to reach semantic scholar. Instead of just failing, leaving a bunch of
pubs at 0, we now have a fallback number in the `citation_count` field
of each publication as well as a `fallback_citation_count_date`. When
a semantic scholar requst works, the citation count is updated and the
date is blanked. When the request fails, the citation count value is
used as is (so the fallback is displayed) and the fallback date is not
removed. The HTML then renders the date as `X citation(s) as of $DATE`

We also include a new piece of data `live_data` on the vue object. When
a semantic scholar request fails this is set to `false` and a new
message about citation data not being live is rendered in the HTML,
below the computed stats.